Feature: Add “frequent” items as a new section to the list
This initially sounded quite useful and not too much work.
In the end it involved a fair few changes  and the usefulness is
somewhat limited.
But out of this there is an interesting quirk that's kind of cool.

A new section “frequent” has been added to the list in between “saved”
and “recent”.
This shows a maximum of two entries.
The most copied (none saved) content
The most pasted (none saved) content (is not the same as above)
Entries from this new list can be used, saved and removed as per current
functionality.

The idea is so that frequently copied or pasted items can be seen and
the user might then want to add these to the “saved” section.

As part of this change the “counts” of number of times pasted and number
of times copied have to be recorded.
These have been added as the second and third column in the list but by
default these are 0 width (i.e. hidden).
With some manual column resizing by the user these columns can be
exposed.
When visible you can then see how many times you have copied and pasted
certain content.
This might be quite cool to see after years of using this app how many
times you have say “pasted your email address”.